CORE-468 Removes configurably the recording of contacts from contacts at enrollment. (#179)
* CORE-468 Removes configurably the recording of contacts from contacts at
enrollment.

There is a new property to control whether to perform the contact
recording during the enrollment for a contact case. If no contact
recording is to be made, this is skipped for contact cases.

The links to the next and previous steps are returned accordingly. An
array of the necessary steps for the breadcrumbs is also supplied with
the EnrollmentDto.

The enrollment is reopened when a contact case is converted to an index
case.

* CORE-468 Deletes the unnecessary endpoint and the link to it

Deletes the endpoint DELETE on /enrollment/completion.

* CORE-468 removed third enrollment step for contact cases

Co-authored-by: VILLAN3LL3 <development@gutfrau.de>